[DLMED] set data type for labels (#547)

+++ examples/classification_3d/densenet_evaluation_dict.py
@@ -38,7 +38,7 @@ def main():
         "/workspace/data/medical/ixi/IXI-T1/IXI585-Guys-1130-T1.nii.gz",
     ]
     # 2 binary labels for gender classification: man and woman
-    labels = np.array([0, 0, 1, 0, 1, 0, 1, 0, 1, 0])
+    labels = np.array([0, 0, 1, 0, 1, 0, 1, 0, 1, 0], dtype=np.int64)
     val_files = [{"img": img, "label": label} for img, label in zip(images, labels)]
 
     # Define transforms for image
